refactor: extract clear-sky modeling into captest.clearsky module
Move pvlib_location, pvlib_system, get_tz_index, and csky out of capdata.py
into a new clearsky module, keeping detect_clearsky in capdata for the
filter_clearsky method. Point io.py and the test suite at the new module.
